refactor(data, util): Move standalone functions to util.
Move wa_response(), upsample() and decimate() functions to util.
