+++ QMigrate/util.py
@@ -11,6 +11,7 @@ import logging
 import time
 
 import numpy as np
+from obspy import Trace
 
 
 log_spacer = "="*110
@@ -177,6 +178,125 @@ def trim2sample(time, sampling_rate):
     """
 
     return int(np.ceil(time * sampling_rate) / sampling_rate * 1000) / 1000
+
+
+def wa_response(convert='DIS2DIS', obspy_def=True):
+    """
+    Generate a Wood Anderson response dictionary.
+
+    Parameters
+    ----------
+    convert : str, optional
+        Type of output to convert between; determines the number of complex
+        zeros used. Options are: 'DIS2DIS', 'VEL2VEL', 'VEL2DIS'
+    obspy_def : bool, optional
+        Use the ObsPy definition of the Wood Anderson response (Default).
+        Otherwise, use the IRIS/SAC definition.
+
+    Returns
+    -------
+    WOODANDERSON : dict
+        Poles, zeros, sensitivity and gain of the Wood-Anderson torsion
+        seismograph.
+
+    """
+
+    if obspy_def:
+        # Create Wood-Anderson response - ObsPy values
+        woodanderson = {"poles": [-6.283185 - 4.712j,
+                                  -6.283185 + 4.712j],
+                        "zeros": [0j],
+                        "sensitivity": 2080,
+                        "gain": 1.}
+    else:
+        # Create Wood Anderson response - different to the ObsPy values
+        # http://www.iris.washington.edu/pipermail/sac-help/2013-March/001430.html
+        woodanderson = {"poles": [-5.49779 + 5.60886j,
+                                  -5.49779 - 5.60886j],
+                        "zeros": [0j],
+                        "sensitivity": 2080,
+                        "gain": 1.}
+
+    if convert in ('DIS2DIS', 'VEL2VEL'):
+        # Add an extra zero to go from disp to disp or vel to vel.
+        woodanderson['zeros'].extend([0j])
+
+    return woodanderson
+
+
+def decimate(trace, sr):
+    """
+    Decimate a trace to achieve the desired sampling rate, sr.
+
+    NOTE: data will be detrended and a cosine taper applied before
+    decimation, in order to avoid edge effects when applying the lowpass
+    filter.
+
+    Parameters:
+    -----------
+    trace : `obspy.Trace` object
+        Trace to be decimated.
+    sr : int
+        Output sampling rate.
+
+    Returns:
+    --------
+    trace : `obspy.Trace` object
+        Decimated trace.
+
+    """
+
+    # Work on a copy of the trace
+    trace = trace.copy()
+
+    # Detrend and apply cosine taper
+    trace.detrend('linear')
+    trace.detrend('demean')
+    trace.taper(type='cosine', max_percentage=0.05)
+
+    # Zero-phase lowpass filter at Nyquist frequency
+    trace.filter("lowpass", freq=float(sr) / 2.000001, corners=2,
+                 zerophase=True)
+    trace.decimate(factor=int(trace.stats.sampling_rate / sr),
+                   strict_length=False, no_filter=True)
+
+    return trace
+
+
+def upsample(trace, upfactor):
+    """
+    Upsample a data stream by a given factor, prior to decimation. The
+    upsampling is done using a linear interpolation.
+
+    Parameters
+    ----------
+    trace : `obspy.Trace` object
+        Trace to be upsampled.
+    upfactor : int
+        Factor by which to upsample the data in trace.
+
+    Returns
+    -------
+    out : `obpsy.Trace` object
+        Upsampled trace.
+
+    """
+
+    data = trace.data
+    dnew = np.zeros(len(data)*upfactor - (upfactor - 1))
+    dnew[::upfactor] = data
+    for i in range(1, upfactor):
+        dnew[i::upfactor] = float(i)/upfactor*data[1:] \
+                        + float(upfactor - i)/upfactor*data[:-1]
+
+    out = Trace()
+    out.data = dnew
+    out.stats = trace.stats
+    out.stats.npts = len(out.data)
+    out.stats.starttime = trace.stats.starttime
+    out.stats.sampling_rate = int(upfactor * trace.stats.sampling_rate)
+
+    return out
 
 
 def timeit(f):
